move `app/api` to `app/backendAPI` for clarity, cus we are gonna expose some “frontend API”
