+++ app/commands/commandBindings/editor.js
@@ -1,6 +1,6 @@
 /* @flow weak */
 import store, { dispatch as $d } from '../../store'
-import api from '../../api'
+import api from '../../backendAPI'
 import * as Pane from '../../components/Pane/actions'
 
 export default {
