+++ app/commands/commandBindings/git.js
@@ -1,6 +1,6 @@
 /* @flow weak */
 import store, { dispatch as $d } from '../../store'
-import api from '../../api'
+import api from '../../backendAPI'
 import * as Git from '../../components/Git/actions'
 import * as Modal from '../../components/Modal/actions'
