+++ app/commands/commandBindings/misc.js
@@ -2,7 +2,7 @@
 import store from '../../store'
 const { getState, dispatch: $d } = store
 
-import api from '../../api'
+import api from '../../backendAPI'
 import * as Modal from '../../components/Modal/actions'
 
 export default {
